+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * Added `route` and `request` to `onPageNotFound` event
     * Added file upload/remove support for `Flex Forms`
     * Added support for `flex-required@: not exists` and `flex-required@: '!exists'` in blueprints
+    * Added `$object->getOriginal()` to get flex objects data before it was modified with `update()`
     * Throwing exceptions from Twig templates fires `onDisplayErrorPage.[code]` event allowing better error pages
 3. [](#bugfix)
     * Fixed escaping in PageIndex::getLevelListing()